xong 2/3 gui_quanlyluong

+++ QuanLyLuongSanPham/src/main/java/model/ChiTietHoaDonBan.java
@@ -12,7 +12,7 @@ import javax.persistence.Table;
 @Table(name = "ChiTietHoaDonBan")
 public class ChiTietHoaDonBan {
 	private double donGia;
-	private double soLuong;
+	private int soLuong;
 
 	@Id
 	@ManyToOne
@@ -32,7 +32,7 @@ public class ChiTietHoaDonBan {
 		super();
 	}
 
-	public ChiTietHoaDonBan(double donGia, double soLuong) {
+	public ChiTietHoaDonBan(double donGia, int soLuong) {
 		super();
 		this.donGia = donGia;
 		this.soLuong = soLuong;
@@ -46,11 +46,11 @@ public class ChiTietHoaDonBan {
 		this.donGia = donGia;
 	}
 
-	public double getSoLuong() {
+	public int getSoLuong() {
 		return soLuong;
 	}
 
-	public void setSoLuong(double soLuong) {
+	public void setSoLuong(int soLuong) {
 		this.soLuong = soLuong;
 	}
